feat(functions): introducte `to` function
The `to` function allows the user to transform an
output of the previous step to a dataobject.
This could be a namedtuple or dataclass.
The `to` function also excepts field names.
To use this field names the defined dataobject should
have default values for its fields.
The `to` function will apply the input to the fields
in the given order of the field names