feat(package): add PipelineMemory
`add_to_memory` enables the user to
add the output of a function to the PipelineMemory.
In steps following later in the `pipeline` the user
cann access the PipelineMemory by `use_memory`.
This approach makes it easier to transfer intermediate results to later steps.
Specially if they are not needed by the steps inbetween